requirements.txt
Pillow == 8.3.1
absl_py == 0.13.0
apex == 0.1
filelock == 3.0.12
lorem == 0.1.1
luigi == 3.0.3
lxml == 4.6.3
nltk == 3.4.5
numpy == 1.20.3
packaging == 21.0
pandas == 0.25.2
py_rouge == 1.1
requests == 2.26.0
sacremoses == 0.0.45
scikit_learn == 0.23.2
tokenizers == 0.8.1rc1
torch == 1.5.1
torchvision == 0.6.1
tqdm == 4.61.2
transformers == 3.0.2